requirements.txt
asterisk-ami==0.1.5
boto3==1.12.39
botocore==1.15.39
paramiko==2.7.1
requests==2.24.0
